requirements.txt
pillow~=11.1.0
opencv-python-headless~=4.11.0.86
Flask==3.1.0
gunicorn==23.0.0
pymupdf
requests